Rewrite circle-packing, again.
